+++ stripe_terminal/android/src/main/kotlin/mek/stripeterminal/mappings/ReaderMappings.kt
@@ -76,6 +76,8 @@ fun DeviceType.toApi(): DeviceTypeApi? {
         DeviceType.ETNA -> DeviceTypeApi.ETNA
         DeviceType.STRIPE_S700 -> DeviceTypeApi.STRIPE_S700
         DeviceType.STRIPE_S700_DEVKIT -> DeviceTypeApi.STRIPE_S700_DEVKIT
+        DeviceType.STRIPE_S710 -> DeviceTypeApi.STRIPE_S710
+        DeviceType.STRIPE_S710_DEVKIT -> DeviceTypeApi.STRIPE_S710_DEVKIT
         DeviceType.UNKNOWN -> null
     }
 }
@@ -220,6 +222,8 @@ fun DeviceTypeApi.toHost(): DeviceType? {
         DeviceTypeApi.ETNA -> DeviceType.ETNA
         DeviceTypeApi.STRIPE_S700 -> DeviceType.STRIPE_S700
         DeviceTypeApi.STRIPE_S700_DEVKIT -> DeviceType.STRIPE_S700_DEVKIT
+        DeviceTypeApi.STRIPE_S710 -> DeviceType.STRIPE_S710
+        DeviceTypeApi.STRIPE_S710_DEVKIT -> DeviceType.STRIPE_S710_DEVKIT
         DeviceTypeApi.APPLE_BUILT_IN -> null
     }
 }