+++ stripe_terminal/example/lib/stripe_api.dart
@@ -10,9 +10,7 @@ class StripeApi {
 
   final _stripe = Stripe(secretKey);
 
-  StripeApi() {
-    assert(secretKey.isNotEmpty, 'Must provide a STRIPE_SECRET_KEY from --dart-define arg');
-  }
+  StripeApi() : assert(secretKey != '', 'Must provide a STRIPE_SECRET_KEY from --dart-define arg');
 
   Future<String> createTerminalConnectionToken() async {
     try {